+++ sawtooth-daml-tp/src/main/java/com/blockchaintp/sawtooth/daml/processor/ContextLedgerState.java
@@ -9,7 +9,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
-package com.blockchaintp.sawtooth.daml.processor.impl;
+package com.blockchaintp.sawtooth.daml.processor;
 
 import static com.blockchaintp.sawtooth.timekeeper.Namespace.TIMEKEEPER_GLOBAL_RECORD;
 
@@ -31,7 +31,6 @@ import java.util.zip.Inflater;
 
 import com.blockchaintp.sawtooth.daml.EventConstants;
 import com.blockchaintp.sawtooth.daml.Namespace;
-import com.blockchaintp.sawtooth.daml.processor.LedgerState;
 import com.blockchaintp.sawtooth.daml.protobuf.ConfigurationEntry;
 import com.blockchaintp.sawtooth.daml.protobuf.ConfigurationMap;
 import com.blockchaintp.sawtooth.daml.protobuf.DamlLogEntryIndex;
@@ -57,7 +56,7 @@ import sawtooth.sdk.processor.exceptions.InvalidTransactionException;
  *
  * @author scealiontach
  */
-public final class DamlLedgerState implements LedgerState {
+public final class ContextLedgerState implements LedgerState {
 
   private static final String TIMEMODEL_CONFIG = "com.blockchaintp.sawtooth.daml.timemodel";
 
@@ -69,7 +68,7 @@ public final class DamlLedgerState implements LedgerState {
 
   private static final int COMPRESS_BUFFER_SIZE = 1024;
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(DamlLedgerState.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(ContextLedgerState.class);
 
   /**
    * The state which this class wraps and delegates to.
@@ -79,7 +78,7 @@ public final class DamlLedgerState implements LedgerState {
   /**
    * @param aState the State class which this object wraps.
    */
-  public DamlLedgerState(final Context aState) {
+  public ContextLedgerState(final Context aState) {
     this.state = aState;
   }
